autoreconnect/requirements.txt
pillow
pyautogui
mss